### 2023-11-16 #149 @paddyroddy
SLEPLET: Slepian Scale-Discretised Wavelets in Python (#148)
Submitting Author: Patrick J. Roddy (@paddyroddy)
Package Name: `SLEPLET`
One-Line Description of Package: Slepian Scale-Discretised Wavelets in Python
Repository Link: https://github.com/astro-informatics/sleplet
Version submitted: [v1.4.4](https://pypi.org/project/sleplet/1.4.4)
Archive: [10.5281/zenodo.7268074](https://doi.org/10.5281/zenodo.7268074)
JOSS DOI: [10.21105/joss.05221](https://joss.theoj.org/papers/10.21105/joss.05221)
Version accepted: TBD

---

`SLEPLET` is a Python package for the construction of Slepian wavelets in the spherical and manifold (via meshes) settings. `SLEPLET` handles any spherical region as well as the general manifold setting. The API is documented and easily extendible, designed in an object-orientated manner. Upon installation, `SLEPLET` comes with two command line interfaces - `sphere` and `mesh` - that allow one to easily generate plots on the sphere and a set of meshes using `plotly`. Whilst these scripts are the primary intended use, `SLEPLET` may be used directly to generate the Slepian coefficients in the spherical/manifold setting and use methods to convert these into real space for visualisation or other intended purposes. The construction of the sifting convolution was required to create Slepian wavelets. As a result, there are also many examples of functions on the sphere in harmonic space (rather than Slepian) that were used to demonstrate its effectiveness. `SLEPLET` has been used in the development of various papers.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

Many fields in science and engineering measure data that inherently live on non-Euclidean geometries, such as the sphere. Techniques developed in the Euclidean setting must be extended to other geometries. Due to recent interest in geometric deep learning, analogues of Euclidean techniques must also handle general manifolds or graphs. Often, data are only observed over partial regions of manifolds, and thus standard whole-manifold techniques may not yield accurate predictions. Slepian wavelets are designed for datasets like these. Slepian wavelets are built upon the eigenfunctions of the Slepian concentration problem of the manifold: a set of bandlimited functions that are maximally concentrated within a given region. Wavelets are constructed through a tiling of the Slepian harmonic line by leveraging the existing scale-discretised framework. Whilst these wavelets were inspired by spherical datasets, like in cosmology, the wavelet construction may be utilised for manifold or graph data.

To the author's knowledge, there is no public software that allows one to compute Slepian wavelets
(or a similar approach) on the sphere or general manifolds/meshes. `SHTools` is a `Python` code used for spherical harmonic transforms, which allows one to compute the Slepian functions of the spherical polar cap. A series of `MATLAB` scripts exist in `slepian_alpha`, which permits the calculation of the Slepian functions on the sphere. However, these scripts are very specialised and hard to generalise.

  - If you made a pre-submission enquiry, please paste the link to the corresponding issue, forum post, or other discussion, or `@tag` the editor you contacted: #148
