+++ .github/workflows/deploy.yml
@@ -13,7 +13,7 @@ jobs:
       id-token: write
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/deployment@2b3c6aedb710b5b43bea7de083eabdb73a408515 # v0
+      - uses: paddyroddy/.github/actions/python/deployment@f55b0d5a35a70b4e5090575c788ee3a7c3d290fb # v0
         with:
           pyproject-toml: ./pyproject.toml
           python-version: "3.11"
