### 2023-11-16 41f9783
Add `pre-commit` to `dev` dependencies
