+++ pyproject.toml
@@ -57,6 +57,7 @@ name = "sleplet"
 optional-dependencies = {dev = [
     "build",
     "mypy",
+    "pre-commit",
     "ruff",
     "tox",
     "twine",
