### 2023-11-16 5b3a9e2
Make `mypy` strict
