### 2023-11-16 035666d
Make `mypy` strict (#313)
