+++ pyproject.toml
@@ -97,11 +97,16 @@ paths.source = [
 ]
 
 [tool.mypy]
+disallow_any_generics = false
+disallow_subclassing_any = false
+disallow_untyped_decorators = false
 explicit_package_bases = true
 plugins = [
     "numpy.typing.mypy_plugin",
     "pydantic.mypy",
 ]
+strict = true
+warn_return_any = false
 
 [tool.pytest.ini_options]
 addopts = """
