+++ .github/workflows/deploy.yml
@@ -13,7 +13,7 @@ jobs:
       id-token: write
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/deployment@f55b0d5a35a70b4e5090575c788ee3a7c3d290fb # v0
+      - uses: paddyroddy/.github/actions/python/deployment@ed26e5e29e6e6f95c4298be97823398fdf3fc6bf # v0
         with:
           pyproject-toml: ./pyproject.toml
           python-version: "3.11"
