+++ .github/workflows/licence.yml
@@ -10,7 +10,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/licence@f55b0d5a35a70b4e5090575c788ee3a7c3d290fb # v0
+      - uses: paddyroddy/.github/actions/licence@ed26e5e29e6e6f95c4298be97823398fdf3fc6bf # v0
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           licence-file: ./LICENCE.md
