### 2023-11-17 b1d6493
Add `pre-commit` default options
