+++ pyproject.toml
@@ -101,10 +101,12 @@ disallow_any_generics = false
 disallow_subclassing_any = false
 disallow_untyped_decorators = false
 explicit_package_bases = true
+ignore_missing_imports = true
 plugins = [
     "numpy.typing.mypy_plugin",
     "pydantic.mypy",
 ]
+scripts_are_modules = true
 strict = true
 warn_return_any = false
 
