### 2023-11-17 932d5d8
Fix `matplotlib.pyplot.xaxis` for `mypy`
