+++ examples/wavelets/axisymmetric_tiling.py
@@ -25,7 +25,7 @@ def main() -> None:
         yi = scipy.interpolate.pchip(x, k)
         plt.semilogx(xi, yi(xi), label=rf"$\Psi^{{{j+J_MIN}}}_{{\ell0}}$")
     plt.xlim(1, xlim)
-    ticks = 2 ** np.arange(np.log2(xlim) + 1, dtype=int)
+    ticks = list(2 ** np.arange(np.log2(xlim) + 1, dtype=int))
     plt.xticks(ticks, ticks)
     plt.xlabel(r"$\ell$")
     plt.legend(loc=6)
