### 2023-11-17 c59a391
Remove `scripts_are_modules`
