+++ pyproject.toml
@@ -106,7 +106,6 @@ plugins = [
     "numpy.typing.mypy_plugin",
     "pydantic.mypy",
 ]
-scripts_are_modules = true
 strict = true
 warn_return_any = false
 
