### 2023-11-17 6755f15
Fix running `mypy .` without `pre-commit` (#315)
