+++ pyproject.toml
@@ -106,6 +106,7 @@ disallow_any_generics = false
 disallow_subclassing_any = false
 disallow_untyped_decorators = false
 explicit_package_bases = true
+ignore_missing_imports = true
 plugins = [
     "numpy.typing.mypy_plugin",
     "pydantic.mypy",
