### 2023-11-17 751a9df
Fiy `mypy` for all `dict`
