+++ src/sleplet/plotting/_create_plot_sphere.py
@@ -33,7 +33,7 @@ class PlotSphere:
     _: dataclasses.KW_ONLY
     amplitude: float | None = None
     """Whether to customise the amplitude range of the colour bar."""
-    annotations: list[dict] = pydantic.Field(default_factory=list)
+    annotations: list[dict[str, float | int]] = pydantic.Field(default_factory=list)
     """Whether to display any annotations on the surface plot or not."""
     normalise: bool = True
     """Whether to normalise the plot or not."""
