### 2023-11-17 e0bb7a4
Fix `mypy` for all `PathLike`
