+++ src/sleplet/_data/setup_pooch.py
@@ -16,7 +16,7 @@ _POOCH = pooch.create(
 _POOCH.load_registry_from_doi()
 
 
-def find_on_pooch_then_local(filename: str) -> os.PathLike | None:
+def find_on_pooch_then_local(filename: str) -> os.PathLike[str] | None:
     """Find a file on POOCH first and if not look in data folder."""
     if filename in _POOCH.registry:
         msg = f"Found {filename} at https://doi.org/{_ZENODO_DATA_DOI}"
