### 2023-11-17 bc31169
Add `pytest` to `dev` dependencies
