### 2023-11-17 79aa325
Fix `mypy` for all `npt.NDArray`
