### 2023-11-17 7a81c93
Disable `disallow_untyped_decorators` for `mypy`
