### 2023-11-17 ffea540
Disable `disallow_any_generics` for `mypy` (#316)
