+++ pyproject.toml
@@ -104,7 +104,6 @@ paths.source = [
 
 [tool.mypy]
 disallow_subclassing_any = false
-disallow_untyped_decorators = false
 explicit_package_bases = true
 ignore_missing_imports = true
 plugins = [
