### 2023-11-17 b19e1f4
Revert " Disable `disallow_untyped_decorators` for `mypy` (#317)"
This reverts commit 3947282dacb5c485d32a18f393b82e50c98047b5.
