+++ .github/workflows/deploy.yml
@@ -13,7 +13,7 @@ jobs:
       id-token: write
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/deployment@ed26e5e29e6e6f95c4298be97823398fdf3fc6bf # v0
+      - uses: paddyroddy/.github/actions/python/deployment@c94d3c482e23614da150416f78d829ba9dea1bab # v0
         with:
           pyproject-toml: ./pyproject.toml
           python-version: "3.11"
