+++ .github/workflows/test.yml
@@ -23,7 +23,7 @@ jobs:
           - "3.11"
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/tox@ed26e5e29e6e6f95c4298be97823398fdf3fc6bf # v0
+      - uses: paddyroddy/.github/actions/python/tox@c94d3c482e23614da150416f78d829ba9dea1bab # v0
         with:
           cache-path: |-
             .tox
