+++ tests/conftest.py
@@ -117,17 +117,6 @@ def random_flm() -> npt.NDArray[np.complex_]:
 
 
 @pytest.fixture(scope="session")
-def random_nd_flm() -> npt.NDArray[np.complex_]:
-    """Create multiple random flm."""
-    return np.array(
-        [
-            sleplet.harmonic_methods.compute_random_signal(L, RNG, var_signal=1)
-            for _ in range(ARRAY_DIM)
-        ],
-    )
-
-
-@pytest.fixture(scope="session")
 def mesh() -> sleplet.meshes.Mesh:
     """Create a bird mesh."""
     return sleplet.meshes.Mesh("bird")
