### 2023-11-20 6af1c94
Renovate: paddyroddy/.github (minor) (#319)
[![Mend Renovate logo
banner](https://app.renovatebot.com/images/banner.svg)](https://renovatebot.com)

This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) |
repository | minor | `v0.86.0` -> `v0.93.0` |
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) | action
| digest | `c94d3c4` -> `c6cc281` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any pro
[…]
