+++ .github/workflows/test.yml
@@ -23,7 +23,7 @@ jobs:
           - "3.11"
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/tox@c94d3c482e23614da150416f78d829ba9dea1bab # v0
+      - uses: paddyroddy/.github/actions/python/tox@c6cc281a113dd0d332c90950662e62b23f1a4dad # v0
         with:
           cache-path: |-
             .tox
