### 2023-11-21 6b4216f
Use new `pre-commit` badge (#321)
