+++ .github/workflows/test.yml
@@ -23,7 +23,7 @@ jobs:
           - "3.11"
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/tox@c6cc281a113dd0d332c90950662e62b23f1a4dad # v0
+      - uses: paddyroddy/.github/actions/python/tox@d061da6340db82a52880b8bd9255562adf49de5c # v0
         with:
           cache-path: |-
             .tox
