+++ .github/workflows/paper.yml
@@ -1,3 +1,4 @@
+---
 name: Paper
 
 # yamllint disable-line rule:truthy
