+++ .github/workflows/deploy.yml
@@ -1,3 +1,4 @@
+---
 name: Deploy
 
 # yamllint disable-line rule:truthy
