+++ .github/workflows/test.yml
@@ -1,3 +1,4 @@
+---
 name: Test
 
 # yamllint disable-line rule:truthy
