+++ .github/workflows/licence.yml
@@ -1,3 +1,4 @@
+---
 name: Update Copyright
 
 # yamllint disable-line rule:truthy
