+++ .github/workflows/examples.yml
@@ -1,3 +1,4 @@
+---
 name: Examples
 
 # yamllint disable-line rule:truthy
