+++ .github/workflows/documentation.yml
@@ -1,3 +1,4 @@
+---
 name: Documentation
 
 # yamllint disable-line rule:truthy
