+++ .github/workflows/licence.yml
@@ -1,3 +1,4 @@
+---
 name: Update Copyright
 
 # yamllint disable-line rule:truthy
@@ -10,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/licence@c6cc281a113dd0d332c90950662e62b23f1a4dad # v0
+      - uses: paddyroddy/.github/actions/licence@d061da6340db82a52880b8bd9255562adf49de5c # v0
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           licence-file: ./LICENCE.md
