+++ .github/workflows/deploy.yml
@@ -1,3 +1,4 @@
+---
 name: Deploy
 
 # yamllint disable-line rule:truthy
@@ -13,7 +14,7 @@ jobs:
       id-token: write
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/deployment@c6cc281a113dd0d332c90950662e62b23f1a4dad # v0
+      - uses: paddyroddy/.github/actions/python/deployment@d061da6340db82a52880b8bd9255562adf49de5c # v0
         with:
           pyproject-toml: ./pyproject.toml
           python-version: "3.11"
