### 2023-11-21 a6dffdc
Rely on config from `pre-commit` hook (#322)
