+++ pyproject.toml
@@ -177,9 +177,6 @@ local_scheme = "no-local-version"
 write_to = "src/sleplet/_version.py"
 
 [tool.tomlsort]
-all = true
-spaces_indent_inline_array = 4
-trailing_comma_inline_array = true
 overrides."project.classifiers".inline_arrays = false
 overrides."tool.coverage.paths.source".inline_arrays = false
 overrides."tool.ruff.isort.section-order".inline_arrays = false
