### 2023-11-21 1280c6a
Remove logo colour
