+++ .github/workflows/test.yml
@@ -24,7 +24,7 @@ jobs:
           - "3.11"
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/tox@d061da6340db82a52880b8bd9255562adf49de5c # v0
+      - uses: paddyroddy/.github/actions/python/tox@931cbf5aea2aceae87072826c7053fa2c6491911 # v0
         with:
           cache-path: |-
             .tox
