### 2023-11-21 20971d7
Change to `LICENCE.txt`
