### 2023-11-21 #149 @NickleDave
Hi @paddyroddy I'm adding the initial Editor-in-Chief checks.

The tl;dr is that we are about ready to start but I need you to do two things first:
- [ ] Add a code of conduct
- [ ] Add a "layperson" summary of the package to the docs page
  - The first thing someone should read on the docs is a statement of what the package does--I wouldn't assume someone will bother to click again to go to PyPI. Provide enough background that someone outside your domain will have some understanding of how the package fits into a bigger picture, kind of like the first couple of sentences of the abstract of your thesis.

I will add one more comment below but it is not required that you address it before we start the review

---
---

- when I first tested that I could import the package, it downloaded a lot of data. Many people might find this surprising. I would suggest adding data directly to the package if possible--if the files are small enough--and/or deferring download of large files
