### 2023-11-21 3735d70
Change to `LICENCE.txt` (#324)
