### 2023-11-21 #149 @NickleDave
@paddyroddy related to that to-do item for the docs:
my understanding is that the goal for SLEPLET is to provide general access to the wavelet basis/transform you have developed for incomplete data on manifolds (feel free to correct me if I'm wrong)

Part of what we can do with this review is help you reach a broader audience. But that means we need to know where to look for editors and reviewers.

Can you please help me understand the main domains where you have applied and can apply your methods? I see you are an RSE in an astro group and that you mention cosmology in your thesis; in the papers you apply your methods to topographic data from Earth. Are these methods mainly used by people in cosmology, geophysics, computer imaging, all of the above?

As examples of places where we might look for editors/reviewers: we work with [astropy](https://www.astropy.org/about.html) and [software underground](https://softwareunderground.org/). I can also imagine asking people that work on manifolds in general and geometric deep learning, and another place we might look would be users/developers of packages you rely on or mention, e.g. pyssht/py2let. But my sense is that we will want domain experts that can help you translate the methods you've developed for a broader audience.

Any more concrete detail you can give me about domains to look at for editors + reviewers would be helpful. We are looking now, but I want to get your input to make sure this review helps you.
