### 2023-11-21 41e64fa
Add `CODE_OF_CONDUCT`
