+++ .github/ISSUE_TEMPLATE/question.yml
@@ -1,0 +1,27 @@
+---
+name: Question
+description: General Questions About Using SLEPLET
+title: "[QUESTION]: "
+labels:
+  - question
+assignees: paddyroddy
+body:
+  - id: topic
+    type: dropdown
+    attributes:
+      label: What is the Topic of Your Question
+      description: Please indicate the topic in the title of your question.
+      options:
+        - Installation
+        - Usage
+        - Documentation
+        - Other
+    validations:
+      required: true
+  - id: question
+    type: textarea
+    attributes:
+      label: Add Your Question Below
+      placeholder: Write your question here
+    validations:
+      required: true
