+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,45 +1,0 @@
----
-name: Bug Report
-about: Create a Report to Help us Improve
-title: "[BUG]"
-labels: bug
-assignees: paddyroddy
----
-
-**Describe the Bug**
-A clear and concise description of what the bug is.
-
-**To Reproduce**
-A minimal working example of code to reproduce the unexpected behaviour.
-
-```python
-import sleplet
-
-...
-```
-
-**Expected Behaviour**
-A clear and concise description of what you expected to happen.
-
-**Actual Behaviour**
-Be a specific and detailed as you can. Paste any output or stack traces of
-errors you receive.
-
-**Version In Use**
-Can be found by any of the following:
-
-```python
-import sleplet
-
-print(sleplet.__version__)
-```
-
-```sh
-mesh -v
-sphere -v
-```
-
-**Additional Context**
-
-- Which version of `Python` are you using?
-- Which OS are you using?
