+++ paper/paper.md
@@ -17,54 +17,89 @@ date: "`r format(Sys.time(), '%d %B %Y')`"
 bibliography: paper.bib
 ---
 
-# Summary
+# Summary <!-- markdownlint-disable-line MD025 -->
 
-Wavelets are widely used in various disciplines to analyse signals both in space and scale.
-Whilst many fields measure data on manifolds (i.e., the sphere), often data are only observed on a partial region of the manifold.
-Wavelets are a typical approach to data of this form, but the wavelet coefficients that overlap with the boundary become contaminated and must be removed for accurate analysis.
-Another approach is to estimate the region of missing data and to use existing whole-manifold methods for analysis.
-However, both approaches introduce uncertainty into any analysis.
-Slepian wavelets enable one to work directly with only the data present, thus avoiding the problems discussed above.
-Applications of Slepian wavelets to areas of research measuring data on the partial sphere include gravitational/magnetic fields in geodesy, ground-based measurements in astronomy, measurements of whole-planet properties in planetary science, geomagnetism of the Earth, and cosmic microwave background analyses.
+Wavelets are widely used in various disciplines to analyse signals both in space
+and scale. Whilst many fields measure data on manifolds (i.e., the sphere),
+often data are only observed on a partial region of the manifold. Wavelets are a
+typical approach to data of this form, but the wavelet coefficients that overlap
+with the boundary become contaminated and must be removed for accurate analysis.
+Another approach is to estimate the region of missing data and to use existing
+whole-manifold methods for analysis. However, both approaches introduce
+uncertainty into any analysis. Slepian wavelets enable one to work directly with
+only the data present, thus avoiding the problems discussed above. Applications
+of Slepian wavelets to areas of research measuring data on the partial sphere
+include gravitational/magnetic fields in geodesy, ground-based measurements in
+astronomy, measurements of whole-planet properties in planetary science,
+geomagnetism of the Earth, and cosmic microwave background analyses.
 
-# Statement of Need
+# Statement of Need <!-- markdownlint-disable-line MD025 -->
 
-Many fields in science and engineering measure data that inherently live on non-Euclidean geometries, such as the sphere.
-Techniques developed in the Euclidean setting must be extended to other geometries.
-Due to recent interest in geometric deep learning, analogues of Euclidean techniques must also handle general manifolds or graphs.
-Often, data are only observed over partial regions of manifolds, and thus standard whole-manifold techniques may not yield accurate predictions.
-Slepian wavelets are designed for datasets like these.
-Slepian wavelets are built upon the eigenfunctions of the Slepian concentration problem of the manifold [@Slepian1961; @Landau1961; @Landau1962]: a set of bandlimited functions that are maximally concentrated within a given region.
-Wavelets are constructed through a tiling of the Slepian harmonic line by leveraging the existing scale-discretised framework [@Wiaux2008; @Leistedt2013].
-Whilst these wavelets were inspired by spherical datasets, like in cosmology, the wavelet construction may be utilised for manifold or graph data.
+Many fields in science and engineering measure data that inherently live on
+non-Euclidean geometries, such as the sphere. Techniques developed in the
+Euclidean setting must be extended to other geometries. Due to recent interest
+in geometric deep learning, analogues of Euclidean techniques must also handle
+general manifolds or graphs. Often, data are only observed over partial regions
+of manifolds, and thus standard whole-manifold techniques may not yield accurate
+predictions. Slepian wavelets are designed for datasets like these. Slepian
+wavelets are built upon the eigenfunctions of the Slepian concentration problem
[…]
