+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,0 +1,65 @@
+---
+name: Bug Report
+description: Create a Report to Help us Improve
+title: "[BUG]: "
+labels:
+  - bug
+assignees: paddyroddy
+body:
+  - id: describe
+    type: textarea
+    attributes:
+      label: Describe the Bug
+      description: A clear and concise description of what the bug is.
+    validations:
+      required: true
+  - id: reproduce
+    type: textarea
+    attributes:
+      label: To Reproduce
+      description: >-
+        A minimal working example of code to reproduce the unexpected behaviour,
+        this will render as `Python` code so no need for backticks.
+      value: |-
+        import sleplet
+
+        ...
+      render: python
+    validations:
+      required: true
+  - id: expected
+    type: textarea
+    attributes:
+      label: Expected Behaviour
+      description: >-
+        A clear and concise description of what you expected to happen.
+    validations:
+      required: true
+  - id: actual
+    type: textarea
+    attributes:
+      label: Actual Behaviour
+      description: >-
+        Be a specific and detailed as you can.
+        Paste any output or stack traces of errors you receive.
+    validations:
+      required: true
+  - id: version
+    type: input
+    attributes:
+      label: Version In Use
+      description: |-
+        Can be found by
+        ```sh
+        python -c "import sleplet; print(sleplet.__version__)"
+        ```
+    validations:
+      required: true
+  - id: additional
+    type: textarea
+    attributes:
+      label: Additional Context
+      value: |-
+        - Python version:
+        - Operating system:
+      render: markdown
