+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -1,0 +1,38 @@
+---
+name: Feature Request
+description: Suggest a Way to Improve This Project
+title: "[FEATURE]: "
+labels:
+  - enhancement
+assignees: paddyroddy
+body:
+  - id: problem
+    type: textarea
+    attributes:
+      label: Is Your Feature Request Related to a Problem? Please Describe
+      description: A clear and concise description of what the problem is.
+      placeholder: I'm always frustrated when [...]
+    validations:
+      required: true
+  - id: solution
+    type: textarea
+    attributes:
+      label: Describe the Solution You'd Like
+      description: A clear and concise description of what you want to happen.
+    validations:
+      required: true
+  - id: alternatives
+    type: textarea
+    attributes:
+      label: Describe Alternatives You've Considered
+      description: >-
+        A clear and concise description of any alternative
+        solutions or features you've considered.
+    validations:
+      required: true
+  - id: additional
+    type: textarea
+    attributes:
+      label: Additional Context
+      description: >-
+        Add any other context or screenshots about the feature request here.
