+++ documentation/DOCUMENTATION.md
@@ -60,9 +60,20 @@ import pyssht as ssht
 import sleplet
 
 for ell in range(2, 0, -1):
-    f = sleplet.functions.HarmonicGaussian(128, l_sigma=10**ell, m_sigma=10)
-    flm = f.translate(alpha=0.75 * np.pi, beta=0.125 * np.pi)
-    f_sphere = ssht.inverse(flm, f.L, Method="MWSS")
+    f = sleplet.functions.HarmonicGaussian(
+        128,
+        l_sigma=10**ell,
+        m_sigma=10,
+    )
+    flm = f.translate(
+        alpha=0.75 * np.pi,
+        beta=0.125 * np.pi,
+    )
+    f_sphere = ssht.inverse(
+        flm,
+        f.L,
+        Method="MWSS",
+    )
     sleplet.plotting.PlotSphere(
         f_sphere,
         f.L,
@@ -81,10 +92,23 @@ sphere earth -L 128
 import pyssht as ssht
 import sleplet
 
-f = sleplet.functions.Earth(128)
-flm = sleplet.harmonic_methods.rotate_earth_to_south_america(f.coefficients, f.L)
-f_sphere = ssht.inverse(flm, f.L, Method="MWSS")
-sleplet.plotting.PlotSphere(f_sphere, f.L, "fig_2").execute()
+f = sleplet.functions.Earth(
+    128,
+)
+flm = sleplet.harmonic_methods.rotate_earth_to_south_america(
+    f.coefficients,
+    f.L,
+)
+f_sphere = ssht.inverse(
+    flm,
+    f.L,
+    Method="MWSS",
+)
+sleplet.plotting.PlotSphere(
+    f_sphere,
+    f.L,
+    "fig_2",
+).execute()
 ```
 
 #### Sifting Convolution on the Sphere: Fig. 3
@@ -100,12 +124,32 @@ import pyssht as ssht
 import sleplet
 
 for ell in range(2, 0, -1):
-    f = sleplet.functions.HarmonicGaussian(128, l_sigma=10**ell, m_sigma=10)
-    g = sleplet.functions.Earth(128)
-    flm = f.convolve(f.coefficients, g.coefficients.conj())
-    flm_rot = sleplet.harmonic_methods.rotate_earth_to_south_america(flm, f.L)
-    f_sphere = ssht.inverse(flm_rot, f.L, Method="MWSS")
-    sleplet.plotting.PlotSphere(f_sphere, f.L, f"fig_3_ell_{ell}").execute()
+    f = sleplet.functions.HarmonicGaussian(
+        128,
+        l_sigma=10**ell,
+        m_sigma=10,
+    )
+    g = sleplet.functions.Earth(
+        128,
+    )
+    flm = f.convolve(
+        f.coefficients,
+        g.coefficients.conj(),
+    )
+    flm_rot = sleplet.harmonic_methods.rotate_earth_to_south_america(
+        flm,
+        f.L,
+    )
+    f_sphere = ssht.inverse(
+        flm_rot,
+        f.L,
+        Method="MWSS",
+    )
+    sleplet.plotting.PlotSphere(
+        f_sphere,
+        f.L,
+        f"fig_3_ell_{ell}",
+    ).execute()
 ```
 
 ### Slepian Scale-Discretised Wavelets on the Sphere
@@ -133,14 +177,39 @@ import pyssht as ssht
 import sleplet
 
 # a
-f = sleplet.functions.Earth(128, smoothing=2)
-flm = sleplet.harmonic_methods.rotate_earth_to_south_america(f.coefficients, f.L)
-f_sphere = ssht.inverse(flm, f.L, Method="MWSS")
[…]
