### 2023-11-22 54e3a73
Only exclude with `MANIFEST.in` (#331)
