+++ .github/workflows/linting.yml
@@ -14,6 +14,6 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/linting@c2013abbbd54e9ea46017879588e2e915afd087c # v0
+      - uses: paddyroddy/.github/actions/linting@37294f6997fb9f62afa155f09808dc3c34459fb2 # v0
         with:
           pre-commit-config: ./.pre-commit-config.yaml
