+++ .github/workflows/test.yml
@@ -24,7 +24,7 @@ jobs:
           - "3.11"
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/tox@c2013abbbd54e9ea46017879588e2e915afd087c # v0
+      - uses: paddyroddy/.github/actions/python/tox@37294f6997fb9f62afa155f09808dc3c34459fb2 # v0
         with:
           cache-path: |-
             .tox
