+++ .github/workflows/deploy.yml
@@ -14,7 +14,7 @@ jobs:
       id-token: write
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/deployment@c2013abbbd54e9ea46017879588e2e915afd087c # v0
+      - uses: paddyroddy/.github/actions/python/deployment@37294f6997fb9f62afa155f09808dc3c34459fb2 # v0
         with:
           pyproject-toml: ./pyproject.toml
           python-version: "3.11"
