+++ .github/workflows/licence.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/licence@37294f6997fb9f62afa155f09808dc3c34459fb2 # v0
+      - uses: paddyroddy/.github/actions/licence@5aa2f0fea5f954ca1f8146d44e9c618c99eb64ca # v0
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           licence-file: ./LICENCE.txt
