### 2023-11-23 ebbe0c9
Renovate: paddyroddy/.github (minor) (#332)
[![Mend Renovate logo
banner](https://app.renovatebot.com/images/banner.svg)](https://renovatebot.com)

This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) |
repository | minor | `v0.110.0` -> `v0.111.0` |
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) | action
| digest | `37294f6` -> `5aa2f0f` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any p
[…]
