### 2023-11-23 829a82c
Add short intro to documentation
