+++ .github/workflows/examples_skip.yml
@@ -1,0 +1,31 @@
+---
+name: Examples
+
+# yamllint disable-line rule:truthy
+on:
+  push:
+    branches:
+      - main
+      - "renovate/**"
+    paths-ignore:
+      - ".github/workflows/**"
+      - "examples/**"
+      - "pyproject.toml"
+      - "src/**"
+  pull_request:
+    paths-ignore:
+      - ".github/workflows/**"
+      - "examples/**"
+      - "pyproject.toml"
+      - "src/**"
+
+jobs:
+  codeblocks:
+    runs-on: ubuntu-latest
+    steps:
+      - run: echo codeblocks
+
+  examples:
+    runs-on: ubuntu-latest
+    steps:
+      - run: echo examples
