### 2023-11-24 #149 @paddyroddy
> The tl;dr is that we are about ready to start but I need you to do two things first:
>
>     * [ ]  Add a code of conduct
>
>     * [ ]  Add a "layperson" summary of the package to the docs page
>
>       * The first thing someone should read on the docs is a statement of what the package does--I wouldn't assume someone will bother to click again to go to PyPI. Provide enough background that someone outside your domain will have some understanding of how the package fits into a bigger picture, kind of like the first couple of sentences of the abstract of your thesis.

I've added these. Mine isn't particularly lay, but they are a fairly specialised application, so hopefully from the description it's obvious who the target audience are!
