### 2023-11-24 #149 @paddyroddy
> ## Editor comments
>
>     * when I first tested that I could import the package, it downloaded a lot of data. Many people might find this surprising. I would suggest adding data directly to the package if possible--if the files are small enough--and/or deferring download of large files

Annoyingly, I hadn't spotted this, as I cache after the first import. I've had a look and the following are created:
```
-rw-------    1 paddy staff  74M Nov 24 18:37 EGM2008_Topography_flms_L2190.mat
-rw-------    1 paddy staff 5.6K Nov 24 18:37 meshes_laplacians_basis_functions_bird_b697_eigenvalues.npy
-rw-------    1 paddy staff  15M Nov 24 18:37 meshes_laplacians_basis_functions_bird_b697_eigenvectors.npy
-rw-------    1 paddy staff 5.6K Nov 24 18:37 meshes_laplacians_slepian_functions_bird_b697_N194_eigenvalues.npy
-rw-------    1 paddy staff 1.1M Nov 24 18:37 meshes_laplacians_slepian_functions_bird_b697_N194_eigenvectors.npy
-rw-------    1 paddy staff  132 Nov 24 18:37 slepian_masks_africa_L1.npy
-rw-------    1 paddy staff  132 Nov 24 18:37 slepian_masks_south_america_L1.npy
```
The problem (in particular) is PyPI doesn't allow over 50MB. I need to work out a way to delay the downloading of `EGM2008_Topography_flms_L2190.mat` in particular.
