+++ src/sleplet/functions/axisymmetric_wavelet_coefficients_africa.py
@@ -29,8 +29,8 @@ class AxisymmetricWaveletCoefficientsAfrica(Flm):
     j: int | None = None
     """Option to select a given wavelet. `None` indicates the scaling function,
     whereas `0` would correspond to the selected `j_min`."""
-    _africa: Africa = pydantic.Field(
-        default=Africa(1),
+    _africa: Africa | None = pydantic.Field(
+        default=None,
         init_var=False,
         repr=False,
     )
