### 2023-11-24 #149 @paddyroddy
> The problem (in particular) is PyPI doesn't allow over 50MB. I need to work out a way to delay the downloading of `EGM2008_Topography_flms_L2190.mat`.

Fixed this in astro-informatics/sleplet#336, turns out I had introduced the bug using `pydantic.Field(default=` values. The data will still be downloaded as and when - but not on initial import. And, as before, it is cached so will only happen once.
