+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -28,8 +28,6 @@ body:
       description: >-
         A clear and concise description of any alternative
         solutions or features you've considered.
-    validations:
-      required: true
   - id: additional
     type: textarea
     attributes:
