### 2023-11-24 86a56e5
Simplify disabling `MD025`
