### 2023-11-24 1f9a510
Simplify disabling `markdownlint` (#338)
