### 2023-11-28 a18557f
Add decorator function to speed up import
