+++ src/sleplet/_data/setup_pooch.py
@@ -1,5 +1,6 @@
 import logging
 import os
+import typing
 
 import platformdirs
 import pooch
@@ -8,19 +9,56 @@ _logger = logging.getLogger(__name__)
 
 
 _ZENODO_DATA_DOI = "10.5281/zenodo.7767698"
-_POOCH = pooch.create(
-    path=pooch.os_cache("sleplet"),
-    base_url=f"doi:{_ZENODO_DATA_DOI}/",
-    registry=None,
-)
-_POOCH.load_registry_from_doi()
+_POOCH = None
 
 
+def _lazy_load_registry(
+    func: typing.Callable[..., os.PathLike[str] | None]
+) -> typing.Callable[..., os.PathLike[str] | None]:
+    """
+    Lazily loads POOCH registry before executing a function.
+
+    Args:
+        func: Function to be decorated
+
+    Returns:
+        Decorated function
+    """
+
+    def wrapper(
+        *args: typing.Any,  # noqa: ANN401
+        **kwargs: typing.Any,  # noqa: ANN401
+    ) -> os.PathLike[str] | None:
+        """
+        Load the POOCH registry if not already loaded and execute the function.
+
+        Args:
+            *args: Positional arguments for the function
+            **kwargs: Keyword arguments for the function.
+
+        Returns:
+            Result of the decorated function
+        """
+        global _POOCH  # noqa: PLW0603
+        if _POOCH is None:
+            _POOCH = pooch.create(
+                path=pooch.os_cache("sleplet"),
+                base_url=f"doi:{_ZENODO_DATA_DOI}/",
+                registry=None,
+            )
+            _POOCH.load_registry_from_doi()
+        return func(*args, **kwargs)
+
+    return wrapper
+
+
+@_lazy_load_registry
 def find_on_pooch_then_local(filename: str) -> os.PathLike[str] | None:
     """Find a file on POOCH first and if not look in data folder."""
-    if filename in _POOCH.registry:
-        msg = f"Found {filename} at https://doi.org/{_ZENODO_DATA_DOI}"
-        _logger.info(msg)
+    if _POOCH is not None:
+        if filename in _POOCH.registry:
+            msg = f"Found {filename} at https://doi.org/{_ZENODO_DATA_DOI}"
+            _logger.info(msg)
         return _POOCH.fetch(filename, progressbar=True)
     if (platformdirs.user_data_path() / filename).exists():
         msg = f"Found {filename} at {platformdirs.user_data_path() / filename}"
