+++ src/sleplet/_data/setup_pooch.py
@@ -60,10 +60,12 @@ def find_on_pooch_then_local(filename: str) -> os.PathLike[str] | None:
             msg = f"Found {filename} at https://doi.org/{_ZENODO_DATA_DOI}"
             _logger.info(msg)
         return _POOCH.fetch(filename, progressbar=True)
+
     if (platformdirs.user_data_path() / filename).exists():
         msg = f"Found {filename} at {platformdirs.user_data_path() / filename}"
         _logger.info(msg)
         return platformdirs.user_data_path() / filename
+
     msg = f"No {filename} found, calculating..."
     _logger.info(msg)
     return None
