+++ src/sleplet/_data/setup_pooch.py
@@ -54,7 +54,15 @@ def _lazy_load_registry(
 
 @_lazy_load_registry
 def find_on_pooch_then_local(filename: str) -> os.PathLike[str] | None:
-    """Find a file on POOCH first and if not look in data folder."""
+    """
+    Find a file on POOCH first and if not look in data folder.
+
+    Args:
+        filename: Filename to find
+
+    Returns:
+        The sought after file or nothing if not found
+    """
     if filename in _POOCH.registry:  # type: ignore[union-attr]
         msg = f"Found {filename} at https://doi.org/{_ZENODO_DATA_DOI}"
         _logger.info(msg)
