### 2023-11-28 9745407
Add decorator function to speed up import (#339)
