+++ pyproject.toml
@@ -61,6 +61,7 @@ optional-dependencies = {dev = [
     "pytest",
     "ruff",
     "tox",
+    "tuna",
     "twine",
 ], docs = [
     "pdoc3",
