### 2023-11-28 #149 @NickleDave
> The problem (in particular) is PyPI doesn't allow over 50MB. I need to work out a way to delay the downloading of EGM2008_Topography_flms_L2190.mat.

Sounds like you fixed the issue overall but jic you're still figuring this out: are you familiar with [pooch](https://github.com/fatiando/pooch)?
A pattern a lot of core scientific Python packages seem to be converging on is "put the data in the package and use importlib-resources if it fits on PyPI; for everything else use pooch". See for example scikit-image: https://github.com/scikit-image/scikit-image/issues/5146 or librosa: https://github.com/librosa/librosa/tree/main/librosa/util/example_data
