### 2023-11-28 #149 @NickleDave
Thank you @paddyroddy for your replies about the domain, very helpful.
I am very sympathetic to developing methods that are useful across domains (FFT anyone?) but just want to make sure we're helping you reach that broader audience.

We are looking for an editor + reviewers now.
