### 2023-11-28 #149 @paddyroddy
> Sounds like you fixed the issue overall but jic you're still figuring this out: are you familiar with [pooch](https://github.com/fatiando/pooch)?

Yes, I'm using `pooch`...
