+++ .github/workflows/documentation.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/pdoc@5aa2f0fea5f954ca1f8146d44e9c618c99eb64ca # v0
+      - uses: paddyroddy/.github/actions/python/pdoc@1a3be93a5bbb79658efc285ce792da79745d9d32 # v0
         with:
           docs-dependency-section: .[docs]
           gh-pages-publish-directory: ./html/sleplet
