+++ .github/workflows/licence.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/licence@5aa2f0fea5f954ca1f8146d44e9c618c99eb64ca # v0
+      - uses: paddyroddy/.github/actions/licence@1a3be93a5bbb79658efc285ce792da79745d9d32 # v0
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           licence-file: ./LICENCE.txt
