+++ .github/workflows/deploy.yml
@@ -14,7 +14,7 @@ jobs:
       id-token: write
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/deployment@5aa2f0fea5f954ca1f8146d44e9c618c99eb64ca # v0
+      - uses: paddyroddy/.github/actions/python/deployment@1a3be93a5bbb79658efc285ce792da79745d9d32 # v0
         with:
           pyproject-toml: ./pyproject.toml
           python-version: "3.11"
