+++ .github/workflows/linting.yml
@@ -14,6 +14,6 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/linting@5aa2f0fea5f954ca1f8146d44e9c618c99eb64ca # v0
+      - uses: paddyroddy/.github/actions/linting@1a3be93a5bbb79658efc285ce792da79745d9d32 # v0
         with:
           pre-commit-config: ./.pre-commit-config.yaml
