+++ LICENCE.txt
@@ -1,6 +1,6 @@
 BSD 3-Clause Licence
 
-Copyright (c) 2017-2023, Patrick Roddy
+Copyright (c) 2017-2024, Patrick Roddy
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
