### 2024-01-02 9320de0
Renovate: actions/setup-python to v5 (#341)
[![Mend
Renovate](https://app.renovatebot.com/images/banner.svg)](https://renovatebot.com)

This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
| [actions/setup-python](https://togithub.com/actions/setup-python) |
action | major | `v4` -> `v5` |

---

### Release Notes

<details>
<summary>actions/setup-python (actions/setup-python)</summary>

### [`v5`](https://togithub.com/actions/setup-python/compare/v4...v5)

[Compare
Source](https://togithub.com/actions/setup-python/compare/v4...v5)

</details>

---

##
[…]
