+++ .github/workflows/linting.yml
@@ -14,6 +14,6 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/linting@1a3be93a5bbb79658efc285ce792da79745d9d32 # v0
+      - uses: paddyroddy/.github/actions/linting@fa62b1a34c1455612031d36abf65f9e792966c95 # v0
         with:
           pre-commit-config: ./.pre-commit-config.yaml
