### 2024-01-02 7b2c42e
Renovate: paddyroddy/.github (minor) (#344)
[![Mend
Renovate](https://app.renovatebot.com/images/banner.svg)](https://renovatebot.com)

This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) |
repository | minor | `v0.112.0` -> `v0.123.0` |
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) | action
| digest | `1a3be93` -> `fa62b1a` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any problems
ther
[…]
