### 2024-01-15 7e5f075
Switch one more `if, elif` to `match case`
