### 2024-01-24 ffbb088
Try `numba` release candidate
