### 2024-01-24 4abf262
Revert "Try `numba` release candidate"
This reverts commit ffbb0884b8f4f641d23fc328b3ed542feb1ba767.
