+++ CITATION.cff
@@ -1,16 +1,17 @@
-cff-version: "1.2.0"
+cff-version: 1.2.0
 authors:
   - family-names: Roddy
     given-names: Patrick J.
-    orcid: "https://orcid.org/0000-0002-6271-1700"
+    orcid: https://orcid.org/0000-0002-6271-1700
 doi: 10.5281/zenodo.7268074
-message: If you use this software, please cite our article in the
+message: >-
+  If you use this software, please cite our article in the
   Journal of Open Source Software.
 preferred-citation:
   authors:
     - family-names: Roddy
       given-names: Patrick J.
-      orcid: "https://orcid.org/0000-0002-6271-1700"
+      orcid: https://orcid.org/0000-0002-6271-1700
   date-published: 2023-04-20
   doi: 10.21105/joss.05221
   issn: 2475-9066
@@ -21,6 +22,6 @@ preferred-citation:
   start: 5221
   title: "SLEPLET: Slepian Scale-Discretised Wavelets in Python"
   type: article
-  url: "https://joss.theoj.org/papers/10.21105/joss.05221"
+  url: https://joss.theoj.org/papers/10.21105/joss.05221
   volume: 8
 title: "SLEPLET: Slepian Scale-Discretised Wavelets in Python"
