+++ .github/workflows/paper.yml
@@ -26,7 +26,7 @@ jobs:
       - name: Upload
 
         # yamllint disable-line rule:line-length
-        uses: actions/upload-artifact@c7d193f32edcb7bfad88892161225aeda64e9392 # v4
+        uses: actions/upload-artifact@5d5d22a31266ced268874388b861e4b58bb5c2f3 # v4
         with:
           name: paper
           path: paper/paper.pdf
