+++ .github/workflows/test.yml
@@ -24,7 +24,7 @@ jobs:
           - "3.11"
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/tox@fa62b1a34c1455612031d36abf65f9e792966c95 # v0
+      - uses: paddyroddy/.github/actions/python/tox@5db0473ebf89b7132850d118505c172be8022f97 # v0
         with:
           cache-path: |-
             .tox
