+++ .github/workflows/documentation.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/pdoc@fa62b1a34c1455612031d36abf65f9e792966c95 # v0
+      - uses: paddyroddy/.github/actions/python/pdoc@5db0473ebf89b7132850d118505c172be8022f97 # v0
         with:
           docs-dependency-section: .[docs]
           gh-pages-publish-directory: ./html/sleplet
