+++ pyproject.toml
@@ -139,6 +139,7 @@ ignore = [
     "D205",
     "D212",
     "D407",
+    "D413",
     "ISC001",
     "N803",
     "N806",
