### 2024-02-06 f209e69
Renovate: paddyroddy/.github (minor) (#350)
[![Mend
Renovate](https://app.renovatebot.com/images/banner.svg)](https://renovatebot.com)

This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) |
repository | minor | `v0.123.0` -> `v0.130.0` |
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) | action
| digest | `fa62b1a` -> `5db0473` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any problems
ther
[…]
