+++ .github/workflows/deploy.yml
@@ -14,7 +14,7 @@ jobs:
       id-token: write
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/deployment@fa62b1a34c1455612031d36abf65f9e792966c95 # v0
+      - uses: paddyroddy/.github/actions/python/deployment@5db0473ebf89b7132850d118505c172be8022f97 # v0
         with:
           pyproject-toml: ./pyproject.toml
           python-version: "3.11"
