+++ .github/workflows/linting.yml
@@ -14,6 +14,6 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/linting@fa62b1a34c1455612031d36abf65f9e792966c95 # v0
+      - uses: paddyroddy/.github/actions/linting@5db0473ebf89b7132850d118505c172be8022f97 # v0
         with:
           pre-commit-config: ./.pre-commit-config.yaml
