+++ pyproject.toml
@@ -132,7 +132,7 @@ testpaths = [
 [tool.ruff]
 fix = true
 force-exclude = true
-ignore = [
+lint.ignore = [
     "COM812",
     "D105",
     "D203",
@@ -144,7 +144,7 @@ ignore = [
     "N803",
     "N806",
 ]
-isort = {known-first-party = [
+lint.isort = {known-first-party = [
     "sleplet",
 ], section-order = [
     "future",
@@ -157,7 +157,7 @@ isort = {known-first-party = [
     "pys2let",
     "pyssht",
 ]}}
-per-file-ignores = {"examples*" = [
+lint.per-file-ignores = {"examples*" = [
     "D100",
     "INP001",
     "T201",
@@ -169,11 +169,11 @@ per-file-ignores = {"examples*" = [
     "S101",
     "SLF001",
 ]}
-select = [
+lint.select = [
     "ALL",
 ]
-mccabe.max-complexity = 18
-pep8-naming.classmethod-decorators = [
+lint.mccabe.max-complexity = 18
+lint.pep8-naming.classmethod-decorators = [
     "classmethod",
     "pydantic.field_validator",
 ]
@@ -185,7 +185,7 @@ write_to = "src/sleplet/_version.py"
 [tool.tomlsort]
 overrides."project.classifiers".inline_arrays = false
 overrides."tool.coverage.paths.source".inline_arrays = false
-overrides."tool.ruff.isort.section-order".inline_arrays = false
+overrides."tool.ruff.lint.isort.section-order".inline_arrays = false
 
 [tool.tox]
 legacy_tox_ini = """
