+++ pyproject.toml
@@ -135,12 +135,10 @@ force-exclude = true
 lint.ignore = [
     "COM812",
     "D105",
-    "D203",
     "D205",
     "D212",
     "D407",
     "D413",
-    "ISC001",
     "N803",
     "N806",
 ]
