### 2024-02-14 facbde3
Use `lint` prefix for `ruff` configuration (#351)
