+++ pyproject.toml
@@ -140,6 +140,7 @@ lint.ignore = [
     "D212",
     "D407",
     "D413",
+    "ISC001",
     "N803",
     "N806",
 ]
