+++ src/sleplet/meshes/mesh_harmonic_coefficients.py
@@ -20,9 +20,9 @@ class MeshHarmonicCoefficients(MeshCoefficients):
 
     def _add_noise_to_signal(
         self: typing_extensions.Self,
-    ) -> tuple[npt.NDArray[np.complex_ | np.float_] | None, float | None]:
+    ) -> tuple[npt.NDArray[np.complex128 | np.float64] | None, float | None]:
         """Add Gaussian white noise to the signal."""
-        self.coefficients: npt.NDArray[np.complex_ | np.float_]
+        self.coefficients: npt.NDArray[np.complex128 | np.float64]
         if self.noise is not None:
             unnoised_coefficients = self.coefficients.copy()
             nlm = sleplet.noise._create_mesh_noise(self.coefficients, self.noise)
@@ -34,7 +34,7 @@ class MeshHarmonicCoefficients(MeshCoefficients):
     @abc.abstractmethod
     def _create_coefficients(
         self: typing_extensions.Self,
-    ) -> npt.NDArray[np.complex_ | np.float_]:
+    ) -> npt.NDArray[np.complex128 | np.float64]:
         raise NotImplementedError
 
     @abc.abstractmethod
