### 2024-02-15 7201b46
Bump `ruff` (#354)
