+++ src/sleplet/meshes/mesh_field.py
@@ -19,7 +19,7 @@ class MeshField(MeshHarmonicCoefficients):
 
     def _create_coefficients(
         self: typing_extensions.Self,
-    ) -> npt.NDArray[np.complex_ | np.float_]:
+    ) -> npt.NDArray[np.complex128 | np.float64]:
         """Compute field on the vertices of the mesh."""
         field = igl.per_vertex_normals(self.mesh.vertices, self.mesh.faces)[:, 1]
         return sleplet.harmonic_methods.mesh_forward(self.mesh, field)
