+++ src/sleplet/functions/africa.py
@@ -28,7 +28,7 @@ class Africa(Flm):
 
     def _create_coefficients(
         self: typing_extensions.Self,
-    ) -> npt.NDArray[np.complex_ | np.float_]:
+    ) -> npt.NDArray[np.complex128 | np.float64]:
         return sleplet.harmonic_methods._ensure_f_bandlimited(
             self._grid_fun,
             self.L,
@@ -54,9 +54,9 @@ class Africa(Flm):
 
     def _grid_fun(
         self: typing_extensions.Self,
-        theta: npt.NDArray[np.float_],  # noqa: ARG002
-        phi: npt.NDArray[np.float_],  # noqa: ARG002
-    ) -> npt.NDArray[np.float_]:
+        theta: npt.NDArray[np.float64],  # noqa: ARG002
+        phi: npt.NDArray[np.float64],  # noqa: ARG002
+    ) -> npt.NDArray[np.float64]:
         """Define the function on the grid."""
         earth_flm = sleplet._data.create_earth_flm.create_flm(
             self.L,
