+++ src/sleplet/meshes/mesh_slepian_functions.py
@@ -26,7 +26,7 @@ class MeshSlepianFunctions(MeshSlepianCoefficients):
 
     def _create_coefficients(
         self: typing_extensions.Self,
-    ) -> npt.NDArray[np.complex_ | np.float_]:
+    ) -> npt.NDArray[np.complex128 | np.float64]:
         """Compute field on the vertices of the mesh."""
         msg = (
             f"Slepian eigenvalue {self.rank}: "
