+++ .github/workflows/examples.yml
@@ -1,12 +1,11 @@
 ---
 name: Examples
 
-# yamllint disable-line rule:truthy
 on:
   push:
     branches:
       - main
-      - "renovate/**"
+      - renovate/**
   pull_request:
 
 concurrency:
