+++ .github/workflows/licence.yml
@@ -1,10 +1,9 @@
 ---
 name: Update Copyright
 
-# yamllint disable-line rule:truthy
 on:
   schedule:
-    - cron: "0 3 1 1 *"
+    - cron: 0 3 1 1 *
 
 jobs:
   update-licence-year:
