+++ .github/workflows/deploy.yml
@@ -1,11 +1,10 @@
 ---
 name: Deploy
 
-# yamllint disable-line rule:truthy
 on:
   push:
     tags:
-      - "v*"
+      - v*
 
 jobs:
   deploy:
