+++ .github/workflows/test.yml
@@ -1,12 +1,11 @@
 ---
 name: Test
 
-# yamllint disable-line rule:truthy
 on:
   push:
     branches:
       - main
-      - "renovate/**"
+      - renovate/**
   pull_request:
 
 jobs:
@@ -50,9 +49,8 @@ jobs:
         uses: coverallsapp/github-action@3dfc5567390f6fa9267c0ee9c251e4c8c3f18949 # v2
         with:
           parallel-finished: true
-          carryforward: "\
+          carryforward: \
             run-macos-latest-3.10,\
             run-ubuntu-latest-3.10,\
             run-macos-latest-3.11,\
-            run-ubuntu-latest-3.11\
-            "
+            run-ubuntu-latest-3.11
