+++ .github/workflows/documentation.yml
@@ -1,7 +1,6 @@
 ---
 name: Documentation
 
-# yamllint disable-line rule:truthy
 on:
   push:
     branches:
