### 2024-02-19 610a976
Test with `uv`
https://astral.sh/blog/uv
