### 2024-03-05 96fb91b
Renovate: paddyroddy/.github (minor) (#358)
[![Mend
Renovate](https://app.renovatebot.com/images/banner.svg)](https://renovatebot.com)

This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) |
repository | minor | `v0.136.0` -> `v0.146.0` |
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) | action
| digest | `2f931cd` -> `512a88e` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any problems
ther
[…]
