+++ src/sleplet/functions/noise_earth.py
@@ -1,4 +1,5 @@
 """Contains the `NoiseEarth` class."""
+
 import numpy as np
 import numpy.typing as npt
 import pydantic
