+++ src/sleplet/functions/harmonic_gaussian.py
@@ -1,4 +1,5 @@
 """Contains the `HarmonicGaussian` class."""
+
 import numpy as np
 import numpy.typing as npt
 import pydantic
