+++ src/sleplet/wavelet_methods.py
@@ -1,4 +1,5 @@
 """Methods to work with wavelet and wavelet coefficients."""
+
 import numpy as np
 import numpy.typing as npt
 
