+++ src/sleplet/meshes/mesh.py
@@ -1,4 +1,5 @@
 """Contains the `Mesh` class."""
+
 import dataclasses
 
 import numpy as np
