+++ src/sleplet/meshes/mesh_slepian_wavelet_coefficients.py
@@ -1,4 +1,5 @@
 """Contains the `MeshSlepianWaveletCoefficients` class."""
+
 import logging
 
 import numpy as np
