+++ src/sleplet/functions/spherical_harmonic.py
@@ -1,4 +1,5 @@
 """Contains the `SphericalHarmonic` class."""
+
 import numpy as np
 import numpy.typing as npt
 import pydantic
