+++ src/sleplet/functions/squashed_gaussian.py
@@ -1,4 +1,5 @@
 """Contains the `SquashedGaussian` class."""
+
 import numpy as np
 import numpy.typing as npt
 import pydantic
