### 2024-03-05 7aadfbd
Force bump `pre-commit` hooks (#359)
