+++ src/sleplet/functions/slepian_noise_south_america.py
@@ -1,4 +1,5 @@
 """Contains the `SlepianNoiseSouthAmerica` class."""
+
 import numpy as np
 import numpy.typing as npt
 import pydantic
