+++ src/sleplet/functions/identity.py
@@ -1,4 +1,5 @@
 """Contains the `Identity` class."""
+
 import numpy as np
 import numpy.typing as npt
 import pydantic
