+++ src/sleplet/functions/wmap.py
@@ -1,4 +1,5 @@
 """Contains the `Wmap` class."""
+
 import numpy as np
 import numpy.typing as npt
 import pydantic
