+++ src/sleplet/functions/axisymmetric_wavelet_coefficients_south_america.py
@@ -1,4 +1,5 @@
 """Contains the `AxisymmetricWaveletCoefficientsSouthAmerica` class."""
+
 import logging
 
 import numpy as np
