+++ src/sleplet/functions/south_america.py
@@ -1,4 +1,5 @@
 """Contains the `SouthAmerica` class."""
+
 import numpy as np
 import numpy.typing as npt
 import pydantic
