+++ src/sleplet/functions/slepian_africa.py
@@ -1,4 +1,5 @@
 """Contains the `SlepianAfrica` class."""
+
 import numpy as np
 import numpy.typing as npt
 import pydantic
