+++ src/sleplet/functions/slepian_noise_africa.py
@@ -1,4 +1,5 @@
 """Contains the `SlepianNoiseAfrica` class."""
+
 import numpy as np
 import numpy.typing as npt
 import pydantic
