### 2024-03-20 #149 @magsol
Ok, I've finished my initial review. I have to admit up front: this is largely outside my area of expertise. While I understand the high-level use case (my research involves a lot of manifold learning and walking, and may occasionally reside in non-Euclidean space), the specifics of Slepian wavelets are not familiar to me.

That said, the package itself seems extremely well put together and maintained. I was able to run the examples and tests without any* problems, and the documentation--while clearly provided for people who understand Slepian wavelets--is well organized.

The only checkbox I couldn't verify was the presence of a repostatus badge, though given that the package itself is already two minor versions more recent than this review version, it's pretty safe to say this package is actively maintained.

* The only caveat here was the CLI example in the README--one of the Zenodo downloads failed mid-stream and crashed the whole example. However, there is [already an issue open about this](https://github.com/astro-informatics/sleplet/issues/291), and it seems like a Zenodo problem, so other than investigating other hosting options for the needed example files, it's easy to say this isn't the fault of the package itself.
