### 2024-03-20 e3dec58
Add repostatus.org badge (#360)
Mentioned in review of pyOpenSci/software-submission#149
