### 2024-03-20 #149 @paddyroddy
Thanks @magsol. I wasn't aware of repostatus.org, I have added one now in astro-informatics/sleplet#360
