### 2024-03-27 24b6f60
Fix quoted string error
See adrienverge/yamllint#663
