+++ .github/workflows/test.yml
@@ -49,8 +49,10 @@ jobs:
         uses: coverallsapp/github-action@3dfc5567390f6fa9267c0ee9c251e4c8c3f18949 # v2
         with:
           parallel-finished: true
-          carryforward: \
+          # yamllint disable-line rule:quoted-strings
+          carryforward: "\
             run-macos-latest-3.10,\
             run-ubuntu-latest-3.10,\
             run-macos-latest-3.11,\
-            run-ubuntu-latest-3.11
+            run-ubuntu-latest-3.11\
+            "
