+++ .github/workflows/examples.yml
@@ -28,7 +28,7 @@ jobs:
           cache-dependency-path: pyproject.toml
 
       - name: Cache pooch
-        uses: actions/cache@ab5e6d0c87105b4c9c2047343972218f562e4319 # v4
+        uses: actions/cache@0c45773b623bea8c8e75f6c82b208c3cf94ea4f9 # v4
         with:
           path: ~/.cache/sleplet
           key: readme-${{ hashFiles('pyproject.toml') }}
@@ -53,7 +53,7 @@ jobs:
           cache-dependency-path: pyproject.toml
 
       - name: Cache pooch
-        uses: actions/cache@ab5e6d0c87105b4c9c2047343972218f562e4319 # v4
+        uses: actions/cache@0c45773b623bea8c8e75f6c82b208c3cf94ea4f9 # v4
         with:
           path: ~/.cache/sleplet
           key: examples-${{ hashFiles('pyproject.toml') }}
