+++ .github/workflows/licence.yml
@@ -10,7 +10,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/licence@512a88ef0d35d185588d460030660a184fc2db36 # v0
+      - uses: paddyroddy/.github/actions/licence@be31e8f5758700dafb58c5852530e750f1ea69d8 # v0
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           licence-file: ./LICENCE.txt
