### 2024-04-02 9eb9bd2
Renovate: paddyroddy/.github (minor) (#364)
[![Mend
Renovate](https://app.renovatebot.com/images/banner.svg)](https://renovatebot.com)

This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) |
repository | minor | `v0.148.0` -> `v0.157.0` |
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) | action
| digest | `512a88e` -> `be31e8f` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any problems
ther
[…]
