### 2024-04-15 #149 @gnypit
Dear All,

First and foremost I'd like to apologize for such a late answer. Though
unprofessional and thus with no excuse, it has an explanation. I started
the review in February with papers at @paddyroddy's Research Gate profile.
I've been distracted from March onwards due to a combination of everything
that could happen at the same time. I had to change my job, reorganise my
wedding party and budget for my moving to London for the next academic
year, as part of my family that was supposed to sponsor my postgraduate
degree (and therefore paid a very symbolic alimony for multiple years, in
spite of come court orders) started backing out on religious grounds,
pressuring me to cancel my wedding altogether... It's been a very crazy
couple of weeks for me, personally and professionally, taking a toll on my
health as well.

Additionally, I wanted to go back to the sleplet package, as the topic
itself got me interested greatly. Simon informed me that the
functionalities have already been verified, which is great news! Therefore,
for me as a mathematician and physicist from different niches, "user
experience" and clarity of code were priorities. The overall impression is
very good: the code is mainly OOP, with clear organisation of files, their
names, and the names of classes, methods and functions. Though I'm no
expert in this particular field, I was able to find myself in all the
folders and have at least a general idea of what a given file is supposed
to be used for. However, I do have some minor remarks when it comes to
documentation and comments in the code. I might be wrong, as this
repository is dedicated to specialists in a given field - however, from my
personal experience many skilled mathematicians, physicists, and others,
have very basic programming skills. Therefore guides, documentation
<https://github.com/astro-informatics/sleplet/blob/main/documentation/DOCUMENTATION.md>and
step-by-step examples are in my view an absolute necessity and need
expanding. I'm not talking about such elaborate ones, like in the case of
Tensorflow, scikit-learn or Transformers, but rather a simple website, such
as for PyGAD <https://pygad.readthedocs.io/en/latest/> or scipy.stats.
Examples could be, at this stage, in the form of Jupyter notebooks, instead
of scripts with defined functions which I have found so far. It's my
recommendation to create such demonstration notebooks as soon as possible.
Moreover, multiple files require additional comments - even snippets in the
current documentation practically don't have any. I find it a very common
practice, that there are few to no comments in the majority of code and I
always suggest adding more. It minimises time required to understand a
given file, especially during debugging, when one wants to understand a
given part of code without reading the whole file, its dependencies, and so
on.

In general, global and local variables are rarely explained - "SNR" in
SlepianNoiseAfrica in slepian_noise_africa.py
<https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/functions/slepian_noise_africa.py>
is
one of few variables (and acronyms!) that was commented on and defined for
a layman, such as myself 😅 Also, _vars.py
<https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/_vars.py>
in slepian folder is a very short code, but it definitely needs some
explanation. Even if there is an accepted practice to use "phi" symbol for
a given angle, not everybody has to follow it necessarily. Or, if someone
is relatively new to the topic, it makes understanding the code very
cumbersome, due to time and attention required to find some context in
which this or that variable is used, how it's used and what precisely may
it represent. Of course these variables were explained in the current
documentation, but it would be more convenient to copy-paste these into the
proper script or create a link/reference. Even a comment "Explained in the
documentation: >link<" would suffice, come to think of that.

In "tests" codes usually have one short comment for a whole function. Same
goes for files like flm.py
<https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/functions/flm.py>,
fp.py
<https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/functions/fp.py>
and so on - in these cases even the files' names are quite secretive. At
the same time files like noise.py
<https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/noise.py>
have very comprehensive comments in some places - generally I've noticed
that files with "methods", "coefficients" and regarding plotting are the
most clearly written and I'd suggest using them as a standard for the rest
of the library. Including LaTeX formulas (e.g. in squashed_gaussian.py
<https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/functions/squashed_gaussian.py>,
harmonic_gaussian.py
<https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/functions/harmonic_gaussian.py>,
etc.) was a great idea and it'd be great to expand on this while creating
documentation - or at least giving some references, perhaps 🤔 Of course,
some files really are very clear with just two comments, like
_bool_methods.py
<https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/_bool_methods.py>,
but they constitute a minority of the whole package. Some files have
messages in them, such as slepian.py
<https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/functions/slepian.py>,
which increases clarity, but specific comments explaining methods like
_validate_rank in Slepian class or _compute_angles in SlepianDiracDelta
class are in my view preferable.

Finally, some concepts in programming might be very new and difficult to
get used to even for semi-advanced programmers in Python, like
multiprocessing. I've noticed the _parallel_methods.py
<https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/_parallel_methods.py>
using a fork of the multiprocessing library. As there are other libraries
and methods to make parallel operations in Python, also depending on the
hardware one uses (e.g. CUDA), it might be a good idea to significantly
expand examples and explain this part of the library in depth so that if
someone uses this part of the library, it is done quite purposefully.

Overall, the biggest challenge before contributors is to create a friendly
documentation and embark on a boring and tiresome adventure to
add comments. There already are examples of in-depth comments, so there is
a clear and present standard to strive for. Apart from this, the library is
well-organised, written beautifully in OOP in Python, with some functional
parts. I find it intuitive and modern. Also, introducing multiprocessing
shows the author's consideration for computing cost, which is a growing
necessity in today's scientific world - especially for those who have the
skills but cannot afford proper hardware. Should I ever meet someone
looking for tools to work with wavelets

I'll make sure with Simon to put it in a proper place as my input to the
review.

With kind regards,
Jakub T. Gnyp

śr., 20 mar 2024 o 20:29 Simon ***@***.***> napisał(a):

> @magsol <https://github.com/magsol> Thank you for your review! It's great
> to hear that SLEPLET is a well-written package :)
>
> @paddyroddy <https://github.com/paddyroddy> I heard from @gnypit
> <https://github.com/gnypit>, and he will review the package soon.
>
