### 2024-04-15 #149 @SimonMolinsky
Hi @gnypit , sorry to hear you have experienced such things; hopefully, your situation improves!

Thank you for your review; just one more thing: could you copy & paste **the review template** into a comment in this thread and select boxes based on your review: https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template (You should be logged in to GitHub, it won't be possible from email).

@gnypit , @paddyroddy :

```
However, just for your information, the parts that are documentation-light (as you say) are the sections that are private (i.e., not exposed to the users) indicated by the leading _. This does beg the question of who https://github.com/pyOpenSci is aiming at: users, developers, or both?
```

Answering this question: **docs in private functions are not required**. @pyOpenSci focuses on both users and developers - as potential maintainers. Documenting private functions is nice to have but _totally optional_ 😊
