### 2024-04-20 #149 @gnypit
Okay @SimonMolinsky & @paddyroddy - just in case, I'm still dabbling in the functionalities on a Linux VM, but:

## Package Review

- [*] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

#### Documentation

- [*] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README.
- [*] **Installation instructions:** for the development version of the package and any non-standard dependencies in README.
- [*] **Vignette(s)** demonstrating major functionality that runs successfully locally.
- [*] **Function Documentation:** for all user-facing functions.
- [*] **Examples** for all user-facing functions.
- [*] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING.
- [*] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `pyproject.toml` file or elsewhere.

Readme file  requirements

- [*] Package has a README.md file in the root directory.

- [*] The package name
- [*] Badges for:
    - [*] Continuous integration and test coverage,
    - [*] Docs building (if you have a documentation website),
    - [*] A [repostatus.org](https://www.repostatus.org/) badge,
    - [*] Python versions supported,
    - [*] Current package version (on PyPI / Conda).

- [*] Short description of package goals.
- [*] Package installation instructions
- [*] Any additional setup required to use the package (authentication tokens, etc.)
- [*] Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the README.md file.
    - [*] Brief demonstration of package usage (as it makes sense - links to vignettes could also suffice here if package description is clear)
- [*] Link to your documentation website.
- [*] Citation information

#### Usability

- [*] Package documentation is clear and easy to find and use.
- [*] The need for the package is clear
- [*] All functions have documentation and associated examples for use
- [*] The package is easy to install

#### Functionality

- [*] **Installation:** Installation succeeds as documented.
- [*] **Functionality:** Any functional claims of the software been confirmed.
  - [*] Tests cover essential functions of the package and a reasonable range of inputs and conditions.
- [*] **Continuous Integration:** Has continuous integration setup (We suggest using Github actions but any CI platform is acceptable for review)
    A few notable highlights to look at:
    - [*] Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)

- [*] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](http://joss.theoj.org/about#submission_requirements).

- [*] **A short summary** describing the high-level functionality of the software
- [*] **Authors:** A list of authors with their affiliations
- [*] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
- [*] **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

#### Final approval (post-review)

- [*] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

Estimated hours spent reviewing: JOSS time +/- 16h

---

#### Review Comments

The overall impression is very good: the code is mainly OOP, with clear organisation of files, their names, and the names of classes, methods and functions. The documentation and examples could use a little more prose around them or comments within them, as already discussed. However, all user-faced parts of code are very thoroughly written and the private functiones, methods, etc. are clearly indicated. Nevertheless, expanding on comments and guides/tutorials might make it easier for both users and contributors to faster and better orient themselves in the library. It doesn't change the fact, that for specialists in the field who are semi-advanced in Python the clarity of the code, intrinsic messages and equations put in LaTeX, together with the required documentation, seem sufficient.

The one concept in programming might be very new and difficult to get used to even for semi-advanced programmers in Python, like multiprocessing. It concernes the private parts of the code, as the _parallel_methods.py is using a fork of the multiprocessing library. As there are other libraries and methods to make parallel operations in Python, also depending on the hardware one uses (e.g. CUDA), it might be a good idea to significantly expand examples and explain this part of the library in depth so that if someone uses this part of the library, it is done quite purposefully. It is, of course, merely a suggestion on what could be done as a part of future development of the library.

Overall, the biggest challenge before contributors is to make the documentation a little more friendly and embark on a boring and tiresome adventure to add comments or alternatives, as already discussed with author during the review. There already are examples of in-depth comments, so there is a clear and present standard to strive for. Apart from this, the library is very well-organised, written beautifully in OOP in Python, with some functional parts. It is intuitive and modern. Also, introducing multiprocessing shows the author's consideration for computing cost, which is a growing necessity in today's scientific world - especially for those who have the skills but cannot afford proper hardware.

Finally, as a personal comment, though I'm still playing around with the functionalities, as I misunderstood the scope of JOSS review, I can already recommend approving the package and will simply let yopu know here in the comments, whenever and whatever else comes up.
