### 2024-04-23 #149 @paddyroddy
> The one concept in programming might be very new and difficult to get used to even for semi-advanced programmers in Python, like multiprocessing. It concernes the private parts of the code, as the _parallel_methods.py is using a fork of the multiprocessing library. As there are other libraries and methods to make parallel operations in Python, also depending on the hardware one uses (e.g. CUDA), it might be a good idea to significantly expand examples and explain this part of the library in depth so that if someone uses this part of the library, it is done quite purposefully. It is, of course, merely a suggestion on what could be done as a part of future development of the library.

[multiprocess](https://github.com/uqfoundation/multiprocess) is a fork of [multiprocessing](https://github.com/python/cpython/tree/main/Lib/multiprocessing) as you say. The reason for using it is that it uses [dill](https://github.com/uqfoundation/dill) rather than [pickle](https://github.com/python/cpython/blob/main/Lib/pickle.py), which makes it easier to serialise data - and much easier to work with IMO. As mentioned previously, this is in the non-public facing part of the code, so shouldn't be too important for users. However, there is definite scope for speeding up further with `CUDA` etc., depending on what hardware one has available.
