+++ CONTRIBUTING.md
@@ -67,7 +67,8 @@ steps in advance to help us fix any potential bug as fast as possible.
 - Determine if your bug is really a bug and not an error on your side e.g. using
   incompatible environment components/versions (Make sure that you have read the
   [documentation](https://astro-informatics.github.io/sleplet). If you are
-  looking for support, you might want to check [this section](#i-have-a-question)).
+  looking for support, you might want to check
+  [this section](#i-have-a-question)).
 - To see if other users have experienced (and potentially already solved) the
   same issue you are having, check if there is not already a bug report existing
   for your bug or error in the
@@ -87,19 +88,20 @@ steps in advance to help us fix any potential bug as fast as possible.
 
 > You must never report security related issues, vulnerabilities or bugs
 > including sensitive information to the issue tracker, or elsewhere in public.
-> Instead, sensitive bugs must be sent by email to <patrickjamesroddy@gmail.com>.
+> Instead, sensitive bugs must be sent by email to
+> <patrickjamesroddy@gmail.com>.
 
 We use GitHub issues to track bugs and errors. If you run into an issue with the
 project:
 
 - Open an [Issue](https://github.com/astro-informatics/sleplet/issues/new).
-  (Since we can't be sure at this point whether it is a bug or not, we ask you not
-  to talk about a bug yet and not to label the issue.)
+  (Since we can't be sure at this point whether it is a bug or not, we ask you
+  not to talk about a bug yet and not to label the issue.)
 - Explain the behaviour you would expect and the actual behaviour.
 - Please provide as much context as possible and describe the _reproduction_
   _steps_ that someone else can follow to recreate the issue on their own. This
-  usually includes your code. For good bug reports you should isolate the problem
-  and create a reduced test case.
+  usually includes your code. For good bug reports you should isolate the
+  problem and create a reduced test case.
 - Provide the information you collected in the previous section.
 
 Once it's filed:
@@ -107,18 +109,18 @@ Once it's filed:
 - The project team will label the issue accordingly.
 - A team member will try to reproduce the issue with your provided steps. If
   there are no reproduction steps or no obvious way to reproduce the issue, the
-  team will ask you for those steps and mark the issue as `needs-repro`. Bugs with
-  the `needs-repro` tag will not be addressed until they are reproduced.
+  team will ask you for those steps and mark the issue as `needs-repro`. Bugs
+  with the `needs-repro` tag will not be addressed until they are reproduced.
 - If the team is able to reproduce the issue, it will be marked `needs-fix`, as
-  well as possibly other tags (such as `critical`), and the issue will be left to
-  be implemented by someone.
+  well as possibly other tags (such as `critical`), and the issue will be left
+  to be implemented by someone.
 
 ### Suggesting Enhancements
 
-This section guides you through submitting an enhancement suggestion for `SLEPLET`,
-**including completely new features and minor improvements to existing**
-**functionality**. Following these guidelines will help maintainers and the
-community to understand your suggestion and find related suggestions.
+This section guides you through submitting an enhancement suggestion for
+`SLEPLET`, **including completely new features and minor improvements to
+existing** **functionality**. Following these guidelines will help maintainers
+and the community to understand your suggestion and find related suggestions.
 
 #### Before Submitting an Enhancement
 
@@ -140,12 +142,13 @@ community to understand your suggestion and find related suggestions.
 Enhancement suggestions are tracked as
 [GitHub issues](https://github.com/astro-informatics/sleplet/issues).
 
-- Use a **clear and descriptive title** for the issue to identify the suggestion.
+- Use a **clear and descriptive title** for the issue to identify the
+  suggestion.
 - Provide a **step-by-step description of the suggested enhancement** in as many
   details as possible.
[…]
