+++ .github/ISSUE_TEMPLATE/documentation.yaml
@@ -25,5 +25,5 @@ body:
     attributes:
       label: How to Improve This Section
       description: >-
-        Do you have any specific suggestions we
-        could take to improve the documentation?
+        Do you have any specific suggestions we could take to improve the
+        documentation?
