### 2024-04-23 b7f8d66
Renovate: paddyroddy/.github (minor) (#366)
[![Mend
Renovate](https://app.renovatebot.com/images/banner.svg)](https://renovatebot.com)

This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) |
repository | minor | `v0.157.0` -> `v0.163.0` |
| [paddyroddy/.github](https://togithub.com/paddyroddy/.github) | action
| digest | `be31e8f` -> `f8793be` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any problems
ther
[…]
