+++ .github/workflows/test.yaml
@@ -23,7 +23,7 @@ jobs:
           - "3.11"
     steps:
       # yamllint disable-line rule:line-length
-      - uses: paddyroddy/.github/actions/python/tox@be31e8f5758700dafb58c5852530e750f1ea69d8 # v0
+      - uses: paddyroddy/.github/actions/python/tox@f8793becd1b97caa5b40ebc7556b21d9a272d700 # v0
         with:
           cache-path: |-
             .tox
