+++ .github/ISSUE_TEMPLATE/feature_request.yaml
@@ -26,8 +26,8 @@ body:
     attributes:
       label: Describe Alternatives You've Considered
       description: >-
-        A clear and concise description of any alternative
-        solutions or features you've considered.
+        A clear and concise description of any alternative solutions or features
+        you've considered.
   - id: additional
     type: textarea
     attributes:
