### 2024-04-28 #149 @SimonMolinsky
@gnypit thank you for your review! As I understand, we're waiting for the performance checks from your side. If you need any support with the point `Packaging guidelines: The package conforms to the pyOpenSci [packaging guidelines](https://www.pyopensci.org/python-package-guide).` then let me know!
