+++ .github/workflows/paper.yaml
@@ -25,7 +25,7 @@ jobs:
       - name: Upload
 
         # yamllint disable-line rule:line-length
-        uses: actions/upload-artifact@5d5d22a31266ced268874388b861e4b58bb5c2f3 # v4
+        uses: actions/upload-artifact@65462800fd760344b1a7b4382951275a0abb4808 # v4
         with:
           name: paper
           path: paper/paper.pdf
