### 2024-05-21 #149 @paddyroddy
> @paddyroddy coming back to the multiprocessing, my starting the subject was motivated only by giving some suggestions. Of course it's a private part of the code, so it's a secondary matter for now.

Thank you for your comments and review @gnypit!
