### 2024-05-21 #149 @paddyroddy
> @paddyroddy Congratulations, SLEPLET has become officially a part of the pyOpenSci family! Thank you for submitting your package here. I will let you know when it will be updated on the pyOS website as soon as it happens :)

Great, thank you. Excited to have this recognition!
