### 2024-06-03 #149 @SimonMolinsky
@paddyroddy Thank you again. We are wrapping things up to close the review officially. I've seen on PyPI that you added the pyOS badge to your package. **Please fill out the the [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7)** in your spare time :)

And again, you might be interested in **writing a blog post about your package**. Here are examples and the template:

[pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation)
[movingpandas](https://www.pyopensci.org/blog/movingpandas-movement-data-analysis)

TEMPLATE: [here](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) is a markdown example that you could use as a guide when creating your post.

Last but not least, we are growing fast and looking for editors, reviewers, and experts in their fields. Your knowledge could be invaluable to the organization. So **let me know if you would like to join the pyOpenSci community!** :slightly_smiling_face:

And something that does not require your input: I've opened a thread about your package on the [pyOpenSci Discourse](https://pyopensci.discourse.group/t/welcome-sleplet-to-the-pyopensci-ecosystem/481). Your package will soon be visible on the pyOpenSci website, Pull Request with metadata is pending :hourglass:

---

🎉 SLEPLET has been approved by pyOpenSci! Thank you @paddyroddy for submitting SLEPLET, and many thanks to @magsol and @gnypit  for reviewing this package! 😸

- [x] Add the badge for pyOpenSci peer-review to the README.md of SLEPLET. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-submission/issues/149)`.

---
