### 2024-06-03 #149 @paddyroddy
@SimonMolinsky I've made a PR for the blog post https://github.com/pyOpenSci/pyopensci.github.io/pull/412
